requirements.txt.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
gunicorn==20.1.0
psycopg2-binary==2.9.7
requests==2.31.0
pandas==2.0.3
pytz==2023.3
python-dotenv==1.0.0
apscheduler==3.10.4